pyproject.toml
[project]
name = "self-interferometry"
version = "0.1.0"
description = "Self-Mixing Interferometry for Rapid Vibration/Acoustic Spectrograms"
authors = [
    {name = "Nolan Peard", email = "nolan.peard@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy==1.26.4",
    "matplotlib",
    "numba",
    "torch",
    "lightning",
    "h5py",
    "tqdm",
    "pyyaml",
    "scipy",
    "neuraloperator",
]

[project.urls]
"Homepage" = "https://github.com/npeard/self-interferometry"
"Bug Tracker" = "https://github.com/npeard/self-interferometry/issues"

[build-system]
requires = ["flit_core>=3.9.0,<4"]
build-backend = "flit_core.buildapi"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--verbose"

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-cov>=3.0",
    "pytest-xdist>=2.5",
    "pytest-benchmark",
]
dev = [
    "pytest>=8.0.0",
    "black",
    "ruff",
    "codespell",
    "taskipy",
    "pre-commit",
    "nbstripout",
]

[tool.flit.module]
name = "self_interferometry"

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'

[tool.ruff.format]
quote-style = "single"
docstring-code-format = true
skip-magic-trailing-comma = true

[tool.ruff.lint]
select = [
    "F", "E", "W", "C90", "I", "D", "UP", "YTT", "ANN", "BLE", "B", "A", "C4", "FA",
    "INP", "NPY201", "PIE", "T20", "PYI", "PT", "RSE", "RET", "SLF", "SIM", "INT",
    "ARG", "PTH", "PL", "TRY", "FLY", "NPY", "RUF", "S", "N", "PERF", "FURB",
]
extend-select = ["D204", "D400", "D404", "D406", "D410"]
ignore = [
    # Import-related (keep for __init__.py files)
    "F403", "F405",
    # Style preferences
    "E741",  # Ambiguous variable names (sometimes necessary in scientific code)
    # Documentation (reasonable for scientific/research code)
    "D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107", "D205", "D405", "D417",
    # Type annotations (reasonable choices)
    "ANN002", "ANN003", "ANN401",
    # Argument usage (common in scientific APIs)
    "ARG001",
    # Return style preference
    "RET505",
    # Exception handling (reasonable for research code)
    "TRY003",
    # Complexity (reasonable limits for scientific code)
    "PLR0913", "PLR2004",
    # Private access (sometimes needed for scientific libraries)
    "SLF001",
    # Security (may need adjustment based on your needs)
    "S101",  # Use of assert (common in scientific code)
    "S311",  # Random generators (often fine for scientific use)
    # Naming (scientific computing often uses uppercase for matrices/operators)
    "N806", "N802", # Variable names should be lowercase (conflicts with math notation)
]

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.ruff.lint.flake8-annotations]
suppress-none-returning = true

[tool.ruff.lint.flake8-comprehensions]
allow-dict-calls-with-keyword-arguments = true

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"docs/**.py" = ["INP001", "ARG001", "ANN"]
"conftest.py" = ["ANN"]
"tests/**.py" = ["ANN"]
"*.ipynb" = ["ALL"]


[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.codespell]
skip = ".git,*.ipynb"

[tool.taskipy.tasks]
test = "pytest"
lint = "ruff check ."
format = "ruff format . && ruff check --fix ."
spell = "codespell ."
precommit = "pre-commit run --all-files"
all = "task format && task lint && task test"
